third_party/blink/web_tests/fast/events/keyboardevent-code-expected.txt

Tests KeyboardEvent.code attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

AlphaNumeric Keys
PASS lastKeyboardDownEvent.code is 'KeyA'
PASS lastKeyboardUpEvent.code is 'KeyA'
PASS lastKeyboardPressEvent.code is 'KeyA'
PASS lastKeyboardDownEvent.code is 'KeyA'
PASS lastKeyboardDownEvent.code is 'KeyM'
PASS lastKeyboardDownEvent.code is 'KeyM'
PASS lastKeyboardDownEvent.code is 'KeyZ'
PASS lastKeyboardDownEvent.code is 'KeyZ'
PASS lastKeyboardDownEvent.code is 'Digit0'
PASS lastKeyboardUpEvent.code is 'Digit0'
PASS lastKeyboardPressEvent.code is 'Digit0'
PASS lastKeyboardDownEvent.code is 'Digit4'
PASS lastKeyboardDownEvent.code is 'Digit9'
PASS lastKeyboardDownEvent.code is 'Backspace'
Functional Keys
PASS lastKeyboardDownEvent.code is 'ControlLeft'
PASS lastKeyboardUpEvent.code is 'ControlLeft'
PASS lastKeyboardDownEvent.code is 'ControlRight'
PASS lastKeyboardDownEvent.code is 'ShiftLeft'
PASS lastKeyboardDownEvent.code is 'ShiftRight'
PASS lastKeyboardDownEvent.code is 'AltLeft'
PASS lastKeyboardDownEvent.code is 'AltRight'
PASS lastKeyboardDownEvent.code is 'Enter'
PASS lastKeyboardDownEvent.code is 'ContextMenu'
PASS lastKeyboardDownEvent.code is 'Space'
PASS lastKeyboardUpEvent.code is 'Space'
PASS lastKeyboardPressEvent.code is 'Space'
PASS lastKeyboardDownEvent.code is 'Tab'
PASS lastKeyboardUpEvent.code is 'Tab'
Control Pad
PASS lastKeyboardDownEvent.code is 'Insert'
PASS lastKeyboardUpEvent.code is 'Insert'
PASS lastKeyboardDownEvent.code is 'Delete'
PASS lastKeyboardDownEvent.code is 'Home'
PASS lastKeyboardDownEvent.code is 'End'
PASS lastKeyboardDownEvent.code is 'PageUp'
PASS lastKeyboardDownEvent.code is 'PageDown'
PASS lastKeyboardUpEvent.code is 'PageDown'
Arrow Pad Section
PASS lastKeyboardDownEvent.code is 'ArrowRight'
PASS lastKeyboardUpEvent.code is 'ArrowRight'
PASS lastKeyboardDownEvent.code is 'ArrowDown'
PASS lastKeyboardDownEvent.code is 'ArrowLeft'
PASS lastKeyboardDownEvent.code is 'ArrowUp'
PASS lastKeyboardUpEvent.code is 'ArrowUp'
Number Pad Section
PASS lastKeyboardDownEvent.code is 'NumLock'
PASS lastKeyboardUpEvent.code is 'NumLock'
Function Section
PASS lastKeyboardDownEvent.code is 'Escape'
PASS lastKeyboardUpEvent.code is 'Escape'
PASS lastKeyboardDownEvent.code is 'F1'
PASS lastKeyboardDownEvent.code is 'F6'
PASS lastKeyboardDownEvent.code is 'F12'
PASS lastKeyboardDownEvent.code is 'PrintScreen'
PASS lastKeyboardUpEvent.code is 'PrintScreen'
PASS successfullyParsed is true

TEST COMPLETE

- 
